cambios + registros base de datos

+++ main/src/main/java/com/proyecto/ColectivosBuilder.java
@@ -5,13 +5,13 @@ import org.json.simple.JSONObject;
 public class ColectivosBuilder {
     public Colectivos builderColectivos(JSONObject object){
         JSONObject colectivos = (JSONObject) object.get("colectivos");
-        int Linea = (int) colectivos.get("Linea");
-        int Unidades = (int) colectivos.get("Unidades");
+        Long Linea = (Long) colectivos.get("Linea");
+        Long Unidades = (Long) colectivos.get("Unidades");
         Long IngresoMensual = (Long) colectivos.get("IngresoMensual");
         Long Mantenimiento = (Long) colectivos.get("Mantenimiento");
 
         
-        Colectivos resultColectivos = new Colectivos(Linea, Unidades, IngresoMensual,Mantenimiento);
+        Colectivos resultColectivos = new Colectivos(Linea.intValue(), Unidades.intValue(), IngresoMensual,Mantenimiento);
         return resultColectivos;
     }
 }
